+++ prediction_market_agent_tooling/tools/web3_utils.py
@@ -18,7 +18,6 @@ from web3.constants import HASH_ZERO
 from web3.contract.contract import ContractFunction as Web3ContractFunction
 from web3.types import AccessList, AccessListEntry, Nonce, TxParams, TxReceipt
 
-from prediction_market_agent_tooling.config import RPCConfig
 from prediction_market_agent_tooling.gtypes import (
     ABI,
     ChecksumAddress,
@@ -34,7 +33,6 @@ from prediction_market_agent_tooling.gtypes import (
 )
 from prediction_market_agent_tooling.loggers import logger
 from prediction_market_agent_tooling.tools._generic_value import _GenericValue
-from prediction_market_agent_tooling.tools.contract import contract_implements_function
 
 ONE_NONCE = Nonce(1)
 ONE_XDAI = xDai(1)
@@ -408,59 +406,3 @@ def get_receipt_block_timestamp(receipt_tx: TxReceipt, web3: Web3) -> int:
 
 def is_valid_wei(value: Web3Wei) -> bool:
     return MIN_WEI <= value <= MAX_WEI
-
-
-def is_erc20_contract(address: ChecksumAddress, web3: Web3 | None = None) -> bool:
-    """
-    Checks if the given address is an ERC20-compatible contract.
-
-    It estimates it by looking if it implements symbol, name, decimals, totalSupply, balanceOf.
-    """
-    web3 = web3 or RPCConfig().get_web3()
-    return (
-        contract_implements_function(address, "symbol", web3)
-        and contract_implements_function(address, "name", web3)
-        and contract_implements_function(address, "totalSupply", web3)
-        and contract_implements_function(
-            address, "balanceOf", web3, function_arg_types=["address"]
-        )
-    )
-
-
-def is_nft_contract(address: ChecksumAddress, web3: Web3 | None = None) -> bool:
-    """
-    Checks if the given address is an NFT-compatible contract (ERC721 or ERC1155).
-
-    For ERC721, checks for: ownerOf, balanceOf, transferFrom.
-    For ERC1155, checks for: balanceOf, safeTransferFrom.
-
-    Returns True if either ERC721 or ERC1155 interface is detected.
-    """
-    web3 = web3 or RPCConfig().get_web3()
-    is_erc721 = (
-        contract_implements_function(
-            address, "ownerOf", web3, function_arg_types=["uint256"]
-        )
-        and contract_implements_function(
-            address, "balanceOf", web3, function_arg_types=["address"]
-        )
-        and contract_implements_function(
-            address,
-            "transferFrom",
-            web3,
-            function_arg_types=["address", "address", "uint256"],
-        )
-    )
-    if is_erc721:
-        return True
-    is_erc1155 = contract_implements_function(
-        address, "balanceOf", web3, function_arg_types=["address", "uint256"]
-    ) and contract_implements_function(
-        address,
-        "safeTransferFrom",
-        web3,
-        function_arg_types=["address", "address", "uint256", "uint256", "bytes"],
-    )
-    if is_erc1155:
-        return True
-    return False
