+++ tests/contract_utils.py
@@ -1,8 +1,17 @@
 from typing import Any
 
 import pytest
+from web3 import Web3
 
-from prediction_market_agent_tooling.gtypes import Wei
+from prediction_market_agent_tooling.gtypes import ChecksumAddress, Wei
+from prediction_market_agent_tooling.markets.omen.omen_contracts import (
+    SDAI_CONTRACT_ADDRESS,
+    WRAPPED_XDAI_CONTRACT_ADDRESS,
+)
+from prediction_market_agent_tooling.tools.contract_utils import (
+    is_erc20_contract,
+    is_nft_contract,
+)
 from prediction_market_agent_tooling.tools.web3_utils import parse_function_params
 
 
@@ -46,3 +55,27 @@ def test_parse_function_params(
 ) -> None:
     result = parse_function_params(input_params)
     assert result == expected_output
+
+
+@pytest.mark.parametrize(
+    "addr, is_erc20",
+    [
+        (WRAPPED_XDAI_CONTRACT_ADDRESS, True),
+        (SDAI_CONTRACT_ADDRESS, True),
+        (Web3.to_checksum_address("0x0D7C0Bd4169D090038c6F41CFd066958fe7619D0"), False),
+    ],
+)
+def test_is_erc20_contract(addr: ChecksumAddress, is_erc20: bool) -> None:
+    assert is_erc20_contract(addr) == is_erc20
+
+
+@pytest.mark.parametrize(
+    "addr, is_nft",
+    [
+        (WRAPPED_XDAI_CONTRACT_ADDRESS, False),
+        (SDAI_CONTRACT_ADDRESS, False),
+        (Web3.to_checksum_address("0x0D7C0Bd4169D090038c6F41CFd066958fe7619D0"), True),
+    ],
+)
+def test_is_nft_contract(addr: ChecksumAddress, is_nft: bool) -> None:
+    assert is_nft_contract(addr) == is_nft
